backend/requirements.txt
fastapi==0.104.1
mangum==0.17.0
uvicorn[standard]==0.24.0
python-socketio==5.10.0
aiohttp==3.9.0
redis==5.0.1
httpx==0.25.1
elevenlabs==1.2.0
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
